requirements.txt
# Core dependencies
numpy>=1.21.0
sentence-transformers>=2.2.0
python-dateutil>=2.8.2

# Database and file handling
sqlite3>=3.35.0  # Built-in
python-magic>=0.4.24

# Clustering and ML
scikit-learn>=1.0.0
hdbscan>=0.8.28
umap-learn>=0.5.3

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.2

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.3.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0
